+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog
 
 Release date: ``2021-xx-xx``
 
-- `NXPY- <https://jira.nuxeo.com/browse/NXPY->`__:
+- `NXPY-203 <https://jira.nuxeo.com/browse/NXPY-203>`__: Better support badly cased or unknown Portal SSO digest algorithms
 
 Technical changes
 -----------------